Separación de responsabilidades para el archivo de VerMisJobs.jsx
- Se modificó código en VerMisJobs.jsx de forma que quedará más limpio y sin tantas líneas.
- Se creó el archivo useVerMisJobs.js para insertar allí toda la lógica JavaScript que iba previamente en el JSX.
